rewrote `ItemsNormalizerTest.php`
